better error handling

+++ shutdownGrit.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
 
-kill -TERM $(cat grit.pid)
+if [ -e grit.pid ]
+ then
+  kill -TERM $(cat grit.pid)
+  rm -f grit.pid
+ else
+  echo "Grit wasn't running or there wa an error"
+fi
